requirements.txt
mcp>=1.0.0
google-genai>=0.3.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
websockets>=11.0.0
aiofiles>=23.0.0
mcp>=1.0.0
websockets>=11.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0